add categories and tags

+++ src/plugins/FileBuildPlugin/FileBuildPlugin.ts
@@ -70,6 +70,10 @@ export class FileBuildPlugin implements PluginInterface {
 
   private async migrateAssets() {
     await fs.ensureDir(this.assetDistPath);
-    await fs.copy(this.assetSrcPath, this.assetDistPath);
+    if (!fs.exists(this.assetSrcPath)) {
+      console.log("no /assets folder to migrate...");
+    } else {
+      await fs.copy(this.assetSrcPath, this.assetDistPath);
+    }
   }
 }
